Add email templates for Netlify Identity
These emails are sent by Netlify Identity when e.g. a new user is
invited. By default they contain links to the root of the site which
doesn't work: they need to point to NetlifyCMS, i.e. /admin/ and it
looks like having custom templates is the only way to achieve that.